Refactored

+++ Server/mods/deathmatch/logic/packets/CLightsyncPacket.h
@@ -23,9 +23,9 @@ public:
     ePacketID     GetPacketID() const noexcept { return PACKET_ID_LIGHTSYNC; };
     std::uint32_t GetFlags() const noexcept { return PACKET_LOW_PRIORITY; };
 
-    void          AddPlayer(CPlayer* pPlayer) noexcept { m_players.push_back(pPlayer); }
-    std::uint32_t Count() const noexcept { return m_players.size(); }
-    void          Reset() noexcept { m_players.clear(); }
+    void        AddPlayer(CPlayer* pPlayer) noexcept { m_players.push_back(pPlayer); }
+    std::size_t Count() const noexcept { return m_players.size(); }
+    void        Reset() noexcept { m_players.clear(); }
 
     bool Read(NetBitStreamInterface& BitStream) noexcept;
     bool Write(NetBitStreamInterface& BitStream) const noexcept;
